#62 edited the tests, need more work to pass
For the time being, it is simply running a test to trigger the notification feature, but it is not asserting the values. Espresso views is not able to 'capture' the text of the notification most likely due to the fact that the text in the notification is an external trigger and not on the views of the camera.